drop introduction ("/start") pages and instead treat 'introduction's as an implicit interstitial question that every other question depends on
This simplifies a lot of code that had to choose whether or not to go to the task start page or to a task question. Those links all now go to task questions.

The 'introduction' on projects is now converted into an implicit interstitial question and an ask-first dependency is added to all other questions so that it is necessarily asked first. This occurs in the load_modules management command.

The introduction page had two nav links, one to invite someone else to take over (this is on question pages too) and another one to return a task to the person who invited the user (which is nowhere else).

fixes #78