a Task is now only considered finished if all required questions have a non-null answer
The module finished page now shows when there are no unanswered questions, even if the module isn't "finished" under this new meaning.

The list of questions on the module finished page now shows a check or an ex, with an ex for required questions that have a null answer (i.e. the same set of questions that prevent a module from being finished).

fixes #77