load_modules now scans the modules directory tree recursively
This allows us to add symlinks to external directories containing module YAML files that we do not want to put into this repository.

`modules/local` is added to .gitignore to make it easier to add local modules for testing.

When we write module IDs inside the YAML files (`id` and question `module-id` fields), we use relative paths to make the IDs portable/relocatable.