rename the Submit, Skip, and Clear buttons
"Submit" is now called "Keep" or "Update", depending on whether the entered value matches an existing answer. This text only appears when there is an existing answer, and Submit is confusing for the new file type questions where leaving the upload input empty means to keep the value. (When the user hasn't answered the question at all, it was and remains "Next >".)

"Skip" is now called "Clear" if there is an existing answer, since its function is to set a null answer.

"Clear" is now called "Reset", since its function is to revert the question to a non-answered state. This is perhaps mostly for debugging.