refactor imputed questions so that they are performed once along with finding the next question to answer
Since imputed values should be evaluated using a context that follows the dependency tree, rather than linearly as the questions are listed in the YAML file, it was helpful to combine the next_question and impute routines. Imputed values are now only calculated when all of a question's dependencies are available. User answers that somehow got entered but now their dependencies aren't satisfied are dropped.

That necessitated some changes to make project root modules work. Since projects have an introduction that isn't answered by the user, we have to pretend the _introduction question is answered so that project tasks show up at all (i.e. their questions' dependencies are satisfied).

Additionally, this changes where the start of dependency evaluation is. It used to be with the questions mentioned in the module document outputs. For projects, questions are typically not listed in a document output. But they must be evaluated. So we now start the evaluation at questions that are not the dependents of other questions. That means we will always see every question. And since this routine now also handles the next-question logic, questions no longer need to be mentioned in output documents (or be a dependency of one that is) to be asked of the user.