assign UUIDs to tasks and use it when exporting project data rather than database primary keys
We'll need this during import so that repeated imports of new Tasks do not generate a new Task on each import, since with a UUID we can identify that a Task came from a previous import. We can't do that with a primary key because we can't ensure that a primary key created on a new Task will match anything in the import data file.