discussion: use css rather than scripting to change the text of the Close & Comment button
When the discussion app is reused in a site that has a Close & Comment button, the text of the button has four states (Close, Re-open, Close & Comment, Re-open & Comment) depending on whether there is user input in the comment textarea and the state of the issue.

Safari has a quirk that if the button's text is changed via $().text() in an event handler triggered by a click on the button, the click event is not fired. This was happening because the click on the button triggers a blur on the textarea which triggers a change event, where we were updating the button's text depending on whether the textarea has input.

Replacing that with CSS classes, which must be implemented by the containing application, fixes it.