re-do the list of questions that show the user context on the question and module-finished pages
They now compute the list through the same mechanism.

Answered questions are listed in chronological order by when they were first answered so that they necessarily show the user's past and future actions no matter where in the history the current page is pointed at (i.e. if they go back to an earlier question, the questions listed before and after that question should be ones the user answered before and after answering the question they are looking at). Previously the current question was always moved to the end of this list -- now it remains in its chronological place.

Future questions are listed in the order they are likely to be asked of the user by using the dependency tree between questions. Previously they were in an arbitrary order.