split the task view into three views: continue (redirects to the next question or task finished view), question (a particular question named in the url), task finished
All so that we can have a permalink to the module finished page. Some untangling was necessary because page permissions required knowing which question was being shown (for discussion guests), which required evaluating the answers to know the next question. By splitting the next-question logic from the question page view, we can evaluate the next question in a view that doesn't have discussions.

fixes #253