discussion comment notification emails now render the comment from markdown to html
When discussion comments generate notification emails, the comment's text is included in the email. The comment body is plain text interpreted as CommonMark. This works fine in the plain text part of the email. But in the HTML part of the email, the plain text goes through simple HTML escaping. Newlines in the comment body become insignificant whitespace, and the lines would all run together. CommonMark characters would appear raw.

The comment body text is now rendered to HTML for the HTML part of the email.

I had to get rid of the curly quotes that were generated around the Notification.description attribute which I don't think made sense anyway. It was interfering with CommonMark rendering if the first character in the comment body was e.g. a "#" for a heading, because then the renderer saw ' "# ', which is not a heading.