unanswered protocol questions would cause template errors
If a template had {{x.y}} and x is a module-type question with a protocol field,
then we don't know what Module will be used to answer the question and we don't
know whether or not 'y' will be a valid attribute. We were returning a value for
'x' that did not have a 'y', and it would yield an AttributeError that would cause
lots of rendering to fail. Now treat unanswered protocol questions as dicts that
infinitely yield an object for whatever is accessed.