authoring tool - question editor
When answering a question in DEBUG mode, if the user has the guidedmodules.module_change permission, and if the module is stored on the local disk, then offer the user a link to edit the question, which shows a modal. On save, the database record for the ModuleQuestion is updated and the Module's spec data is serialized to YAML and written out to disk, overwriting what was there.

module-type questions are not implemeted here yet.