expose ModuleAssets and user-uploaded files for question answers through views that have authorization logic
We were using the default view for django-db-storage, which uses file content hashes for filenames for... security.
That view is now removed so the assets are not served anonymously. New views are created to serve these blobs
that checks that the requesting user has access. The views set appropriate HTTP headers to avoid serving untrusted
resources in a manner that would allow client-side execution of Javascript on our domain (which would be bad).

fixes #252