re-do 5fd699dfd3b9349d5a21b52b038eeb906314a0a2 and (again) 2246a1e2afe8f96b83d948d3c4696df7cd993b81
Links on the project page were not all clickable. The DOM was not allowing <a> tags to wrap forms
or Bootstrap .row elements, probably because .row's have display:table. So the jQuery click handler
was not constructed right for them DOM that the browser actually constructed.

The orignal motivation for the change was that the new <a> tags for the authoring tool were inside
div's with click handlers, and clicking the <a> also triggered the div's click handler which began
a page navigation. So I removed the div's click handler and made it an <a>, because an <a> around
an <a> seemed to work fine, but it didn't (per above).

This simplifies the DOM by removing all of the <a>s and only using div click handlers, and using
stopPropagation on the authoring tool buttons to prevent page navigation from being triggered
by the outer divs.