replace Module.visible with AppInstance.system_app
The Module.visible and Module.superseded_by fields were a part of our original database model where each Module was a top-level entity and as we created new versions of Modules we'd mark old database versions as not visible, meaning not able to be started by and end-user, and we would set superseded_by to point to the newer version of the Module (so we knew why it was no longer visible). This functionality became obsolete when we moved to apps because the compliance store controlled what Modules could be started, not the visible field in the database.

But we also used the Module.visible field to pick out the current version of system Modules, i.e. the account settings and organization profile Modules, and when we moved to apps this functionality broke. The load_modules management command would create a new AppInstance for the system apps on each invocation, even if nothing changed. And it would set the new Module to visible without making the old version not visible. This lead to "get() returned more than one topic -- it returned ___!" errors while trying to get the current system Module for accounts or organizations.

This commit:

* Drops the Module.visible field (we have no use for its data archivally), and its uses.
* Marks the Module.superseded_by field as obsolete.
* Adds a new field, AppSource.system_app, a boolean field that is true just for the current version of the account settings and organization profile apps.
* Revises load_modules to try a compatible update first and if it fails, revert the transaction and try again by creating a new AppInstance.