authoring tool: add a hidden way to force an update if there are incompatible data model changes
In the Javascript console, execute 'authoring_tool_reload_app_force = true' to enable a forced update.

But when forcing an update, add error handling when a question that has already been answered is attempted to be deleted, which can't be performed because of database constraints.

Also catch YAML syntax errors and dependency errors.