re-position the authoring tool icon on project pages again
Because HTML does not permit links within links, I had made started apps/modules <div>s with onclick handlers so that the authoring tool inside it could be a link. But that means you couldn't right-click or middle-click the link to open it in a new browser tab, which was frustrating. So I've moved the authoring tool DOM outside of the app div and turned the div back into an <a>.

Unstarted apps remain divs because they require form POSTs which can't be opened in a new browser tab.