show whether tasks are finished on app pages & revise definition of finished
We had "finished" labels in the row view (for projects that don't have sub-apps) but there was no indication in grid view, so I added a checkmark using the color we use elsewhere to incidate done-ness.

Also changed the meaning of whether a project is finished - now all of its tasks must not just be started, they must also be finished. This required updating how the cache for finished state is cleared since it must be done recursively for tasks when tasks are answers to other tasks.