add whitenoise for static file serving
It's helpful in development to be able to run with DEBUG=False, but then we lose runserver's static file serving. This adds static files support back.

It will also be helpful in deployments, such as Docker (see #355), that don't have an HTTP layer to serve static files.

The "static" key in our environment.json file must be set to a path to hold flattened static files, and `manage.py collectstatic` must be run.