docker: create a `manage` script in /usr/local/bin that starts "python manage.py" and use it instead of 'docke exec ... ./manage.py'
Since a virtualenv is now used, which was added for the CentOS-based image, manage.py can't be called directly from docker exec anymore since it needs the virtualenv to be started first. The first_run and compliance_app commands broke.

This adds a global `manage` script that actives the virtualenv environment and passes its command-line arguments to manage.py.