fixes for resetting a project title
Allow a project title to be reset to the default by setting it to the empty string. This was blocked client side even though it was meant to be working based on the server side.

When the title is changed, call on_answer_changed since Project titles affect cached answers. And... the title itself is cached when it is computed from instance-name and this is a good way to clear the title cache in case it's incorrectly stale.