use a new source for django-notifications-hq that is patched for Django 2.0 compatibility
Pull django-notifications-hq from a GitHub commit. Since we're pulling a(n unreleased) package from a URL,
we can't put it in requirements.in because pip-compile does not support packages specified by URL. So,
instead, I'm keeping the original package in requirements.in so that pip-compile continues to expand out
django-notifications-hq's dependencies into requirements.txt, but I'm adding a script that rewrites
django-notifications-hq's entry in requirements.txt to pull the package from a GitHub URL with a pinned
commit hash, and I've manually added the package hash values that pip will check.