on project pages, display module-set tasks as separate cards rather than bullets within a card for the question
Revise the redirect after starting an app so it can target a particular card by question and task ID, so when starting module-set question apps it can highlight the right task.

Removing some CSS for highlighing finished questions and the next question the user should be working on which we haven't really been using.