test_screenshots: use a throw-away test database rather than using the local database & deleting records at the end
This is generally safer and creates a clean environment so that screenshots
don't leak local state. But also deleting records at the end of the command
is difficult because of foriegn keys with on_delete=PROTECT. Django isn't
smart enough to delete AppSources when there's a ModuleQuestion whose
answer_type_module is set to another Module in that AppSource. Even though
it all is going to be deleted, Django complains about protection.

Since the database is clean, revert the part of an earlier commit where
we only show the app catalog for a particular source while taking screenshots.