small changes to avoid false positives in static code scanning
1) replace md5 with sha1 in a non-security context to avoid code scan false-positive

In discussions, user avatars without profile pictures fall back to CSS using get_avatar_fallback_css.
To generate a random but stable color, we compute a hash over user profile information. A hash
ensures that we get a uniform distribution of colors but also that it's always the same color for the user.

For speed, I used MD5. Code scanners flag this, but it's a false positive since this is not about security.
I'm changing it to SHA1 so scanners don't flag it.

2) use get_random_string() instead of hard-coding a password

The password being replaced is a test case for a failed login --- it wasn't user
credentials to begin with.

3) add absolute paths to subprocess calls

Shell out to pandoc, wkhtmltopdf, and xvfb-run using absolute paths by adding /usr/bin/ to
the start of the command.