cached_state must be cleared if any Task in the same project is modified
That's because some Tasks have templates that reference {{project}} and
can see all data throughout the project, so whenever anything in a project
changes, those Task's cached_state must be cleared. This has long been
a bug but until 66e306c38b2cc0da7132637cbd8917cae932c354 the only information
that was cached was whether a Task was finished and how far along it is,
so we did not notice the bug. Now that we're caching output document
content, it was readily apparent that the cache was not being cleared
when answers were changing.