add missing account/email/email_confirmation.md template
This is used when a user goes to /accounts/email/, which we don't
link to anywhere(!) but a user may end up at after trying to
verify their email address on signup (or perhaps only if the
verification link expired). The user can then re-send a verification
email. This is the template for the email body.

fixes #469