fix control disabling during discussion comment submission and file attachment
When we added comment drafts, we moved half of the work of submitting a comment into first updating the draft, then the draft is published in a second step. We were lacking UI changes during the
AJAX request for updating the draft and our standard AJAX loading modal was harsh for the second half when the request is slow to finish. There was a similar setup for attachments (save draft to
ensure there's a draft to attach something to, then submit attachment).

This applies control disabling across the two AJAX requests (save draft + {submit | attach}) and drops our modal loading indicator.