fix missing parens in project import error handling
If answer validation fails when importing project data, a message is
logged (for showing the user at the end of the import). But the line
was missing parentheses and an unhandled exception was raised instead.

I encountered this while running tests without an internet connection.
Email address validation failed without a DNS server, and instead of
logging the error the exception caused the test to fail with an error.

I'm also adding a test to check that imports are producing the correct
log output.