disable email address deliverability checks during tests
When running without an internet connection, email address deliverability
checks fail because DNS servers are unavailable. This occurs when validating
answers to 'email-address' questions as well as validating email addresses
given by users when sending invitations.

This is now turned off during tests by setting a flag in the settings.