docker: if --image is specified, don't automatically pull
When debugging I like to specify an image that is local and I don't
want to see an error when the image isn't found on the Docker Hub.
Also in the next commit I am going to have the script abort on error,
and I don't want it to abort if I'm using a local image.