docker: use a read-only root filesystem if a remote database is in use
A read-only root filesystem can be used if there is no persistent data stored in the container. For simple testing, we use a Sqlite database stored inside the container --- which requires a read-write filesystem. Although we write other things to the filesystem (we use /tmp in a few places), we don't write anything that needs to be persistent to disk except for the testing database.

Since Sqlite requires a writable journal file in the same directory as the database file, even if we are bind-mounting the Sqlite database to a host path so that it is persistent and not stored in the container, we still can't use a read-only root filesystem unless the directory containing the Sqlite database (or the journal file itself) is bind-mounted to the host too. So, we don't automatically turn on a read-only filesystem in this case, although we could if we get the journal file bind-mounted to the host. We can't move the database to /tmp like we do with environment.json because a disappearing journal file could corrupt the database, and we expect the database to persist across container reboots even if it's not mounted.

To test this, I used `--dburl sqlite:////tmp/db.sqlite` --- i.e. put the database in a writable location on the container.

* If --dburl is given to docker_container_run.sh, then make the container's root filesystem read-only and mount /run and /tmp as tmpfs filesystems, which are cleared out each time the container starts.
* Allow the container to still write local/environment.json by making it actually a symlink to /tmp/environment.json. Since it's written out on container startup, it's fine for it to live in /tmp.
* Since /tmp is cleared out on container startup, move static assets that are generated during the image build from /tmp to the application source directory.
* Since /tmp is cleared out on container startup, move the container non-root user's home directory out of /tmp and into a normal place like /home/ so that things like .bashrc aren't lost. That's helpful when using 'docker exec ... bash' to log into the machine.