replace the Skip submit button with three new emoji-labeled toggle buttons I Don't Know (shrug), Doesn't Apply (raising hand), and I'm Not Sure (thinking)
I Don't Know and Doesn't Apply both store a null value for the answer, as the old Skip button used to do. The difference is only in the value of the new skipped_reason field on answers. When these are chosen, the main form fields are disabled to indicate to the user that they are storing a null. These options are mutually exclusive.

I'm Not Sure applies to regular answers and Doesn't Apply null answers and sets an unsure flag so the user knows they might want to come back to it. The unsure flag is also always set when the user chooses I Don't Know, but it's not represented in the UI because it looks odd/redundant --- i.e. I'm Not Sure and I Don't Know are mutually exclusive.

Form validation is revised a bit. We no longer change the caption of the Next button, when revisiting an answered question, depending on whether the answer is being changed or not, because that got more complicated with the new fields. Also, the Next button would be disabled when invalid input was given, but now that only applies if I Don't Know/Doesn't Apply aren't chosen.

* add skipped_reason field to TaskAnswerHistory which is null, dont-know, or doesnt-apply
* add unsure boolean flag to TaskAnswerHistory which is true if the user probably should revisit the answer later
* replace the Skip submit button with three new toggle buttons I Don't Know, Doesn't Apply, and I'm Not Sure and translate these into the skipped_reason and unsure flags when saving
* when saving an answer, create a new TaskAnswerHistory if the value is the same but the skipped_reason or unsure flags are different
* upgrade to latest EmojiOne which introduces new emojis that we're using
* show the skip reason and unsure flag in the answer history
* restore the state of the new toggle buttons when revisiting an answered question
* show the skipped reason (Don't Know/Doesn't Apply) and the unsure flag on the review page