docker: use supervisord to run the application server and the notification emails process
Our docker container was not running the background process for sending notification emails. This commit
uses supervisord to start the uWSGI application server and the notifications process.

We now get supervisord's output on the container STDOUT ('docker container logs') instead of uWSGI's
output, which was initialization status and a log of requests. So I've added a new script in the
container tail_logs which via 'docker container exec' can be used to see the uWSGI process output.

Supervisor writes logs, which include sub-process STDOUT/STDERR, to /var/log, which is now mounted
as a tmpfs if docker_container_run.sh is using --read-only to make the rest of the filesystem
read-only. In production deployments, /var/log could be mounted as a volume so that the logs are
persistent or accessible to sysadmins.

Added documentation for where the logs are and a suggestion to mount them using a Docker volume.

Added documentation for the UID and GID of the non-root user, which must be set up in the host
environment if the host environment overrides the container user.