docker: simplify logs path
I started seeing a permission denied error opening the log file at /var/log/supervisor/supervisord.log.
The directory was world-writable. I couldn't see a reason why the file could not be created. But
moving the file to /var/log solved it. When running with a read-only root filesystem and a tmpfs
for /var/log, the director starts off empty anyway and we explicitly created /var/log/supervisor.
Moving the logs to right in /var/log simplifies things a bit by not having to worry about whether
the tmpfs mount locations start empty or not.