templates: escape multi-line Jinja2 expressions before CommonMark rendering
When rendering templates, first we render CommonMark to HTML and then we run Jinja2 template directives,
which is how we know we get escaping right on substituted variables (we do HTML escaping). Before doing
CommonMark rendering we 'hide' Jinja2 expressions so that they don't get messed up, e.g. [link]({% insert_url %})
would result in "{% insert_url %}" being URL-encoded and then not being a Jinja2 directive anymore. But
we were only doing this for directives on a single line because the regex was too strict.