allow project admins to delete sub-tasks using a new trash icon on task pages
We have had the ability to delete Projects, which is a database-level cascaded deletion of the Project's
root task. Projects are instantiated compliance apps. So users could delete entire instantiations of
apps. But we didn't have the ability to delete other tasks, i.e. tasks that are within compliance apps.

The use case I'm fixing is when there's a module-set question at the top level of an app, e.g. for
creating many POAMs, and where the module-set's answer module is in the same app. Each new answer
creates e.g. a new instance of the module, a new POAM, but there was no way to remove these sub-tasks
from the top-level app's answers.

Actually we used to have a feature to mark tasks as deleted but we had taken out the UI for it. This commit
replaces that defunct code with new working code.

The new delete trash icon marks the Task as deleted and *also* removes it from the data model of answers
by updating any question whose answer included that Task to have a new answer with the same answer but
the deleted Task removed. In other words, deletion is really "unlinking" from whatever task it might have
been an answer to, plus the setting of a deleted_at flag so that we know why the record is dangling. We
also set the new answer's "method" to be "delete", to distinguish the method that the answer was created
from a normal answering of a question in the UI or import/API. This new value for the method field requires
a metadata-only database migration.

Also fixing a typo from 30048a50 that prevented the scripts in task-base.html from being included in module-finished.html.