statically serve emojione images from vendored sprites
We had been using EmojiOne's Javascript script to create emoji img elements from emoji names. The script didn't
seem to be compatible with IE 11, giving an error about String.includes not being defined. While
investigating, I noticed that it loaded the emoji images from a CDN at run time (contrary to our policy against
embedding third-party resources) and additionally that the CSS that we were loading in fetch-vendor-resources.sh
was actually a 404 page from GitHub (although its absence didn't seem to affect anything).

We use emojis on question pages and in discussion reactions.

This commit removes the EmojiOne script and replaces it by downloading the underlying image files (as sprites,
i.e. large images that hold a grid of emojis) plus a CSS file that maps emoji Unicode code points to coordinates
in the sprite grid. We serve the image sprites and the CSS statically. The question and discussion templates
are revised to load the emojis by applying CSS classes --- the code points for the emojis we use are added
to the DOM so that we do not need to perform a name-to-codepoint lookup.