drop the ModuleAssetPack model and move the fields into AppInstance
The ModuleAssetPack model was used to quickly determine during the loading
of a compliance app if the total set of assets are identical to a previous
set of assets loaded by an app from the same source (i.e. a previous version
of the app). It also was an intermediate table between Module and ModuleAsset,
which was actually confusing because there was at most one ModuleAssetPack for
all of the Modules in an AppInstance, so it should have been tied to the
AppInstance instead of the individual Modules.

Since ModuleAssetPacks didn't hold much information and were 1:1 with
AppInstance, it simplifies the data model to merge the fields into AppInstance.
If an AppInstance has any of the same assets as a previous AppInstance, the
those assets' binary data are still shared with the previous version of the app
throgh the reuse of ModuleAsset instances (which are keyed off of the contents'
hash).

So, previously, Modules across a whole AppSource reused *mappings* from
paths to binary blobs (in addition to reusing ModuleAssets themselves).
Now that information is stored in AppInstances, so the mapping is still
reused within an app but not across apps.