pyyaml 4.1 raised an error during testing of parsing of 'yaml'-format templates
Our output documents support a 'yaml' experimental template format. In this format, we recursively walk the data structure and replace string values by evaluating them as their own output document templates. However, when we run documents in the PARSE_ONLY mode for validating modules, the inner document templates return junk values if validation passes, and if these junk values are not safe for YAML output (i.e. not basic data types), pyyaml 4.1 raises an exception.

The fix is to not render the substituted data structure back to YAML when the caller is just testing that e.g. Jinja2 syntax is correct.