bump dependencies
The 'typing' package dependency disappeared in this last run of requirements_txt_updater.sh (which runs pip-compile), but CircleCI complains that the package is still required by the fs package:

> In --require-hashes mode, all requirements must have their versions pinned with ==. These do not: typing~=3.6; python_version < "3.6" from https://files.pythonhosted.org/packages/4a/bd/eee1157fc2d8514970b345d69cb9975dcd1e42cd7e61146ed841f6e68309/typing-3.6.6-py3-none-any.whl#sha256=57dcf675a99b74d64dacf6fba08fb17cf7e3d5fdff53d4a30ea2a5e7e52543d4 (from fs==2.1.0->-r requirements.txt (line 119))

So I first manually kept it in requirements.txt, but then our requirements_txt_checker.sh running in CircleCI says actually the package has been upgraded, and I copied the output which has the desired requirements.txt lines into our requirements.txt.

There's probably an upstream packaging error or something weird in my local packages or pip-compile.

Also there's a new version of setuptools not yet in the CircleCI image so adding that to our ignore list, and dropping lines in the ignore list that CircleCI has caught up to.