make Python 3.6+ a requirement
When running requirements_txt_updater.sh (pip-compile) under Python 3.6, the `typing` dependency disappears because it is no longer required on Python 3.6 because the module has been moved into the standard library. However, when installing our dependencies on Python <3.5, one of our dependencies has a (conditional) dependency on the `typing` package. Since it's not listed in our `requirements.txt`, pip correctly refuses to install it because it has not been pinned with a hash. The solution is that if we run `requirements_txt_updater.sh` on Python 3.6+, then we must also run `pip install -r requirements.txt` on Python 3.6+.