fix IAM proxy login reload issue
When using an IAM proxy server, the middleware needs to be arranged so that the proxy headers are read and applied *before* we check if the user has permission to view the organization domain they are requesting. We were seeing that a browser reload was necessary after login to make it to the projects page because the / route is permitted regardless of authorization and the user makes it to that page, then the IAM headers were read and applied to the request user, which started a session, and on reload the session would be used and the user would be permitted to go to their projects page.

The middleware is now reversed, so the user's session is created before the site checks if the user has permission to see organization projects.