drop the concept of an unskippable question dependency which prevented questions from being asked and imputed
Questions can depend on the answers to other questions in various ways, including the use of previous answers in question propmpts and the use of previous answers in impute conditions.

A dependency was considered 'unskippable' if the referenced question was used in an impute condition. If a question was skipped (i.e. I Don't Know/It Doesn't Apply/etc. or imputed null) but considered unskippable, then later questions that depend on it were blocked from being asked of the user and from being imputed. The original idea was probably to create a way to force users to answer certain questions before proceeding to other questions.

But this created unexpected behavior: Questions were not presented to the user, or imputed, when impute condition variables were previously imputed as null, leaving those questions in an unanswered state yet not shown to the user to answer.

This meant that the progress bar that shows module completion status could show the module as finished (no more answerable questions!) but also only half done (many unanswered questions!).