Refactor comments in check_version_file.py to explain VERSION fails (#614)
To help us keep version value synchronized between git tag, and
changelog, we use a VERSION file an python script for CircleCI to run.
It can be confusing when this test fails and how to fix it. Added
improvements to check_version_file.py comments and details to ERROR
message.