user profiles are now site-wide rather than unique to each organization the user is a member of
We previously gave each user a profile that was unique to each multi-tenant organization the user was a member of. On each access to user profile information, we had to request profile information for a particular organization.

Now profiles are site-wide. Lots of places that passed organizations around to get the right profile are now simplified.

The migration permanently deletes profiles so that each user only has one profile. Reversing the migration deletes any remaining profiles because they then lack a required association with an Organization needed in the previous database schema.