authoring tool 'new question' broke after 0e9671f8
In 0e9671f8, I removed code that handled an old or unused condition where a single protocol ID on a question could be specified either as a simple string or as a one-item list containing a string. When the authoring tool is used to add a new question to a project-type module, it adds a question with a dummy protocol as a simple string (not as a list). As a result, the app catalog crashed.

This commit changes the authoring tool to set the question's protocol to a list containing the string.

Although in module YAML files on disk the author can give a protocol as a string or as a list of strings, when we load YAML files into the database we alwys convert single strings into one-item lists, so that the protocol value is uniformly a list (see 90930b00). The code that I had deleted in the app catalog doesn't appear to ever have actually been necessary, and the authoring tool was always creating an incorrect value for protocol in the database, but with both errors together it worked.

Fixes #631.