Fix privacy, tos for subdomain removal; add loving security page
The rendering of the static page had special code and a template
to address rendering from the landing page and from an subdomain-ed
organization. This special code was causing rendering of the
privacy and tos page to fail in 0.9.0. That code is now fixed
and tests have been added to check for the privacy static pages.

We have also added the static page for loving assessments.