Add a way to start assessments from command
This is a bit rough, and the code should be tidied up a bit.
I wasn't originally planning on doing a web-based version of this,
or at least I wasn't going to add it to the main branch,
but *this* part seems to be a bit fiddly/delicate when done via code.