Fix '@-mention' in discussions
Using an '@-mention' in the discussion field
was causing the comment submission to fail because
the code block that invited @-mentioned users was
trying to create the invite with a reference to
the organization from multi-tenancy days. Since
the organization no longer had a value, this caused a problem.

The fix was simple enough (once found) to remove the
mention of the organization in the creation of the invitation
since the user does not need to be directed to specific
multi-tenant subdomain.